Control.MonadEnv

Description

User-facing API for MonadEnv, typeclass used to implement an environment

Documentation

class (Num r, Monad e) => MonadEnv e s a r | e -> s a r where

The environment monad TODO: Think about two typeclasses: ContinuousMonadEnv and EpisodicMonadEnv

Minimal complete definition

reset, step

Methods

reset :: e (Initial s)

Any environment must be initialized with reset. This can be used to reset the environment at any time. It's expected that resetting an environment begins a new episode (and can only be called once in a continuous environment).

step :: a -> e (Obs r s)

Step though an environment with an action, run the action in the environment, and return a reward and the new state of the environment.

data Obs r o

An observation of the environment will either show that the environment is done with the episode (yielding Done), that the environment has already Terminated, or will return the reward of the last action performed and the next state TODO: return Terminal (or return ()) on failure

Constructors

Next !r !o

Done !r !(Maybe o)

Terminated

data Initial o

When starting an episode, we want to send an indication that the environment is starting without conflating this type with future steps (in Obs r o)

Constructors

Initial !o

EmptyEpisode

type Reward = Double

A concrete reward signal.