Control.MonadMWCRandom

Description

typeclass to remove extraneous mwc-random functions

Documentation

class Monad m => MonadMWCRandom m where

MonadMWCRandom for public use. FIXME: use with PrimState so that we can use ST

Minimal complete definition

getGen

Methods

getGen :: m GenIO

type MonadMWCRandomIO m = (MonadIO m, MonadMWCRandom m)

A convenience type constraint with MonadMWCRandom and MonadIO.

type MWCRand = MWCRandT Identity

simple type alias for transformer-less variant

newtype MWCRandT m a

a wrapper to share a generator without using reader

Constructors

MWCRandT

Fields

getMWCRandT :: ReaderT GenIO m a

runMWCRand :: MWCRand a -> GenIO -> a

run a transformerless MWC-random Monad

runMWCRandT :: MWCRandT m a -> GenIO -> m a

unwrap MonadMWCRandom

re-exports from System.Random.MWC

type GenIO = Gen (PrimState IO)

A shorter name for PRNG state in the IO monad.

wrappers for System.Random.MWC

uniform :: (MonadIO m, MonadMWCRandom m, Variate a) => m a

uniform referencing MonadMWCRandom's generator

uniformR :: (MonadIO m, MonadMWCRandom m, Variate a) => (a, a) -> m a

uniformR referencing MonadMWCRandom's generator

wrappers for Statistics.Distribution

genContVar :: (MonadIO m, MonadMWCRandom m, ContGen d) => d -> m Double

genContVar referencing MonadMWCRandom's generator

extras

sampleFrom :: (MonadIO m, MonadMWCRandom m) => [Double] -> m (Int, [Double])

Sample a single index from a list of weights, converting the list into a distribution

class Variate a

The class of types for which we can generate uniformly distributed random variates.

The uniform PRNG uses Marsaglia's MWC256 (also known as MWC8222) multiply-with-carry generator, which has a period of 2^8222 and fares well in tests of randomness. It is also extremely fast, between 2 and 3 times faster than the Mersenne Twister.

Note: Marsaglia's PRNG is not known to be cryptographically secure, so you should not use it for cryptographic operations.

Minimal complete definition

uniform, uniformR

_uniform :: (PrimMonad m, Variate a) => Gen (PrimState m) -> m a

a convenience helper to reference the underlying System.Random.MWC function