Data.CartPole

Description

Shared datatypes between Gym environments and the haskell implementation of CartPole.

Documentation

data StateCP

The state of a cart on a pole in a CartPole environment

Constructors

StateCP

Fields

position :: Float

position of the cart on the track

angle :: Float

angle of the pole with the vertical

velocity :: Float

cart velocity

angleRate :: Float

rate of change of the angle

data Action

Cartpole can only go left or right has an action space of "discrete 2" containing {0..n-1}.

FIXME: Migrate this to either a more generic "directions" actions (would need things like "up", "down" versions as well) or a "discrete actions" version. I'm a fan of the former.

Constructors

GoLeft

GoRight

type Event = Event Double StateCP Action

Specific datatype for a CartPole event