Data.Logger

Description

In lieu of a history monad embedded in models maintain a logger monad for easily debugging environments.

Documentation

data Event r o a

Our primary datatype for an event in a trace. Contains the episode number, reward, state, and action taken (in that order). TODO: change the ordering to Event Integer s a r

Constructors

Event Integer r o a

class Monad m => Logger m where

A logging monad, this is seperate from a History monad in that this is intended to be used for debugging and for toggling log information.

FIXME: In reality, this is halfway between commenting/uncommenting print statements, and passing CPP flags. Something should be done about this or a real logging Monad should be brought in.

Minimal complete definition

info, info_, debug, debug_

Methods

info :: Text -> m ()

log at the info level.

info_ :: Text -> Text -> m ()

log at the info level, appending information on the left.

debug :: Text -> m ()

log at the debug level.

debug_ :: Text -> Text -> m ()

log at the debug level, appending information on the left.

newtype NoopLogger m x

A prebuilt type which doesn't actually log anything.

Constructors

NoopLogger

Fields

runNoopLogger :: m x

newtype DebugLogger m x

A prebuilt type that does all levels of logging

Constructors

DebugLogger

Fields

runDebugLogger :: m x