Environments.Bandits

Description

Implementation of an n-armed bandit environment.

FIXME: currently this is only for a 10-armed bandit. This needs to be tied to a config.

Documentation

newtype Environment a

Monad for an n-armed bandit environment

Constructors

Environment

Fields

getEnvironment :: RWST Config (DList Event) () IO a

runEnvironment :: Config -> Environment () -> IO (DList Event)

run an n-armed bandit environment

data Event r o a

Our primary datatype for an event in a trace. Contains the episode number, reward, state, and action taken (in that order). TODO: change the ordering to Event Integer s a r

Constructors

Event Integer r o a

data Action

The slot machine index whose arm will be pulled

mkBandits :: Int -> Int -> Float -> GenIO -> Config

helper function to build a bandits config with normally-distributed reward functions

defaultBandits :: GenIO -> Config

Give the default config of a 10-armed bandit

mkAction :: Int -> Environment Action

Convert an Int to an Action in the bandit environment. Throw if the Int falls out of bounds.