Environments.Gym.ClassicControl.MountainCarContinuousV0

Description

Environment description: > A car is on a one-dimensional track, positioned between two "mountains". > The goal is to drive up the mountain on the right; however, the car's > engine is not strong enough to scale the mountain in a single pass. > Therefore, the only way to succeed is to drive back and forth to build up > momentum. Here, the reward is greater if you spend less energy to reach the > goal. > > MountainCarContinuous-v0 defines "solving" as getting average reward of > 90.0 over 100 consecutive trials. > > This problem was first described by Andrew Moore in his PhD thesis [Moore90]. > > Here, this is the continuous version.

https://gym.openai.com/envs/MountainCarContinuous-v0

Documentation

data Action

Actions a car can perform to get out of it's predicament

Constructors

MoveLeft

DoNothing

MoveRight

type Runner s a x = RunnerT s a IO x

type alias of RunnerT in IO

data State

State of a car stuck between two hills

Constructors

State

Fields

position :: Float

velocity :: Float

type Environment = EnvironmentT IO

Alias to EnvironmentT in IO

type EnvironmentT t = GymEnvironmentT State Action t

Alias to GymEnvironmentT with MountainCarV0 type dependencies

runEnvironment :: Manager -> BaseUrl -> RunnerT State Action IO x

Alias to runEnvironment in IO

runEnvironmentT :: MonadIO t => Manager -> BaseUrl -> RunnerT State Action t x

Alias to runEnvironmentT

runDefaultEnvironment :: RunnerT State Action IO x

Alias to runDefaultEnvironment in IO

runDefaultEnvironmentT :: MonadIO t => RunnerT State Action t x

Alias to runDefaultEnvironmentT