Environments.Gym.ClassicControl.PendulumV0

Description

The inverted pendulum swingup problem is a classic problem in the control literature. In this version of the problem, the pendulum starts in a random position, and the goal is to swing it up so it stays upright.

https://gym.openai.com/envs/Pendulum-v0

Documentation

newtype Action

Force to exert on the pendulum

Constructors

Action

Fields

getAction :: Float

type Runner s a x = RunnerT s a IO x

type alias of RunnerT in IO

data State

State of a PendulumV0 environment FIXME: give these semantics or move to a tuple?

Constructors

State

Fields

cosTheta :: Float

sinTheta :: Float

thetaDot :: Float

type Environment = EnvironmentT IO

Alias to EnvironmentT in IO

type EnvironmentT t = GymEnvironmentT State Action t

Alias to GymEnvironmentT with PendulumV0 type dependencies

runEnvironment :: Manager -> BaseUrl -> RunnerT State Action IO x

Alias to runEnvironment in IO

runEnvironmentT :: MonadIO t => Manager -> BaseUrl -> RunnerT State Action t x

Alias to runEnvironmentT

runDefaultEnvironment :: RunnerT State Action IO x

Alias to runDefaultEnvironment in IO

runDefaultEnvironmentT :: MonadIO t => RunnerT State Action t x

Alias to runDefaultEnvironmentT