Environments.Gym.Internal

Description

Underlying implementation to run a Gym environment using the gym-http-client.

Documentation

newtype GymEnvironmentT s a t x

The gym environment, which is shared through all openai/gym instances

Constructors

GymEnvironmentT

Fields

getEnvironmentT :: RWST GymConfigs (DList (Event Reward s a)) (LastState s) (ClientT t) x

type RunnerT s a t x = Bool -> GymEnvironmentT s a t x -> t (Either ServantError (DList (Event Reward s a)))

type alias for a runEnvironment type which executes an environment action

type Runner s a x = RunnerT s a IO x

type alias of RunnerT in IO

data GymException

Possible errors we might encounter while interacting with our environment

Constructors

UnexpectedServerResponse String

TypeError String

EnvironmentRequiresReset

runEnvironmentT :: forall o a t x. MonadIO t => GymEnv -> Manager -> BaseUrl -> Bool -> GymEnvironmentT o a t x -> t (Either ServantError (DList (Event Reward o a)))

run an effectful gym to completion and return either an error, or the history of the agent. FIXME: move this into a History monad seperate from the Environment so that we don't blow up memory.

runEnvironment :: GymEnv -> Manager -> BaseUrl -> Bool -> GymEnvironment o a x -> IO (Either ServantError (DList (Event Reward o a)))

same as runEnvironmentT, but with IO as the base monad

runDefaultEnvironmentT :: MonadIO t => GymEnv -> Bool -> GymEnvironmentT o a t x -> t (Either ServantError (DList (Event Reward o a)))

same as runEnvironmentT, however use http-client's default manager settings

runDefaultEnvironment :: GymEnv -> Bool -> GymEnvironment o a x -> IO (Either ServantError (DList (Event Reward o a)))

same as runEnvironment, however use http-client's default manager settings

_reset :: (MonadIO t, MonadThrow t, FromJSON o) => GymEnvironmentT o a t (Initial o)

generic rest function which makes a call to the gym and returns the first observation

_step :: (MonadIO t, MonadThrow t, ToJSON a, r ~ Reward, FromJSON o) => a -> GymEnvironmentT o a t (Obs r o)

generic step function which takes a ToJSONable action and returns a reward and a FromJSONable state