Environments.Gym.ToyText.FrozenLakeV0

Description

The agent controls the movement of a character in a grid world. Some tiles of the grid are walkable, and others lead to the agent falling into the water. Additionally, the movement direction of the agent is uncertain and only partially depends on the chosen direction. The agent is rewarded for finding a walkable path to a goal tile.

https://gym.openai.com/envs/FrozenLake-v0

Documentation

type Runner s a x = RunnerT s a IO x

type alias of RunnerT in IO

newtype StateFL

The current position of the agent on the frozen lake

Constructors

Position

Fields

unPosition :: Int

toVector :: StateFL -> Vector Int

Convert StateFL to a computable type

mkStateFL :: MonadThrow m => Int -> m StateFL

Build a FrozenLakeV0 state, throwing if the position is out of bounds.

type Environment = EnvironmentT IO

Alias to EnvironmentT in IO

type EnvironmentT t = GymEnvironmentT StateFL Action t

Alias to GymEnvironmentT with FrozenLakeV0 type dependencies

runEnvironment :: Manager -> BaseUrl -> RunnerT StateFL Action IO x

Alias to runEnvironment in IO

runEnvironmentT :: MonadIO t => Manager -> BaseUrl -> RunnerT StateFL Action t x

Alias to runEnvironmentT

runDefaultEnvironment :: RunnerT StateFL Action IO x

Alias to runDefaultEnvironment in IO

runDefaultEnvironmentT :: MonadIO t => RunnerT StateFL Action t x

Alias to runDefaultEnvironmentT

data Action

Actions that can be performed in FrozenLakeV0

Constructors

Left

Down

Right

Up