Reinforce.Spaces.Action

Description

typeclass for a discrete action space, as well as helper functions

Documentation

class (Bounded a, Enum a) => DiscreteActionSpace a where

Mostly tags around an Enum, but includes information about the size of an action space and is used in helper functions.

Associated Types

type Size a :: Nat

Methods

toAction :: Int -> a

fromAction :: a -> Int

oneHot :: forall a. (KnownNat (Size a), DiscreteActionSpace a) => a -> R (Size a)

one-hot encode a bounded enumerable. Doesn't care if minBound is or 0

oneHot' :: forall a. DiscreteActionSpace a => a -> Vector Double

one-hot encode a bounded enumerable

allActions :: DiscreteActionSpace a => [a]

helper function to get all actions in a discrete action space

randomChoice :: forall m a. (MonadIO m, MonadMWCRandom m, DiscreteActionSpace a) => m a

make a uniform-random selection of an Action in a discrete action space