Reinforce.Spaces.State

Description

Documentation

class StateSpace s where

State space information to convert to and from a Data.Vector

Minimal complete definition

toVector, fromVector

Methods

toVector :: s -> Vector Double

fromVector :: MonadThrow m => Vector Double -> m s

class StateSpaceStatic s where

State space information to convert to and from a static hmatrix vector

Minimal complete definition

toR

Associated Types

type Size s :: Nat

Methods

toR :: s -> R (Size s)